+++ internal/openvpn/main.go
@@ -21,6 +21,8 @@ import (
 	"github.com/jkroepke/openvpn-auth-oauth2/internal/utils"
 )
 
+const minManagementInterfaceVersion = 5
+
 func New(ctx context.Context, logger *slog.Logger, conf config.Config, oauth2Client *oauth2.Provider) *Client {
 	client := &Client{
 		conf:   conf,
@@ -137,7 +139,7 @@ func (c *Client) checkManagementInterfaceVersion() error {
 
 	// Management Interface Version 5 is required at minimum
 	// ref: https://github.com/OpenVPN/openvpn/commit/a261e173341f8e68505a6ab5a413d09b0797a459
-	if managementInterfaceVersion < 5 {
+	if managementInterfaceVersion < minManagementInterfaceVersion {
 		return errors.New("openvpn-auth-oauth2 requires OpenVPN management interface version 5 or higher")
 	}
